Added a MIN_DISTANCE parameter
Binary search is used to assess whether prototype/signature lines are
too close, if they are, we continue pulling numbers out of the hat until
they aren't. Must ensure distance isn't too high or we'll end up in a
loop forever, may want a check for that. Noise lines are also subject to
this.