Force resizing at 16:10 aspect ratio
Ideally this would be a customisable part of the library, but for now
this seems the best way to get the behaviour I want. If you take a look
in the PsychoJS source code, it actually shrinks the content based on
whatever the smallest dimension is:

    minSize = Math.min(win.size[0], win.size[1]);

This essentially does the kind of resizing I want, at a 1:1 aspect
ratio. You can make it 16:10 by applying a regex that matches with:

    let minSize=Math.min\(([^,]*),([^\)]*)\);

And replaces with (a minified version of):

    let minSize = Math.min($1, $2);
    const a = $1 / $2;
    const req_a = 16/10;
    if (a < req_a) {
        minSize = $1 / req_a;
    }

This is what this commit is. Yeah, best way would be to maintain my own
copy of the PsychoJS source and maybe make a pull request, maybe later.